Add files via upload
Refer to 2a. Image Restricted Configuration in the readme.txt,  we are suggested to use pairs.txt. So that I used the pairsDevTest.txt and pairsDevTrain.txt. I divided the pairsDevTest.txt into the validatation and test dataset. Finally, there are 2200 pairs in train dataset, 500 in validation and 500 in test dataset.